Break up testCMCSingleMuscle into two tests.
These tests will be run in parallel now. However, they previously wrote to the
same files. To get the tests to run in parallel and pass, I had to rename the
directories to which forward and cmc tools wrote, and also the location of the
forward kinemtaics for use in cmc.

Addresses #31.
